backend/pyproject.toml
[project]
name = "parliament-backend"
version = "0.1.0"
description = "Parliament debate backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0,<0.105.0",
    "uvicorn[standard]>=0.24.0,<0.25.0",
    "langchain>=0.1.0,<0.2.0",
    "langchain-google-vertexai>=0.0.1,<0.1.0",
    "google-cloud-aiplatform>=1.38.0,<1.39.0",
    "pydantic>=2.5.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "python-multipart>=0.0.6,<0.1.0",
    "sse-starlette>=1.8.0,<1.9.0",
    "setuptools>=80.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

[tool.hatch.build.targets.wheel]
packages = ["."] 

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D203",  # no-blank-line-before-class
    "D213",  # multi-line-summary-second-line
    "COM812",  # missing-trailing-comma
]
